+++ docker-compose.yaml
@@ -26,8 +26,8 @@ services:
     container_name: mongodb
     ports:
       - "27017:27017"
-    volumes:
-      - ./data/db:/data/db
+    # volumes:
+      # - ./data/db:/data/db
     networks:
       - my_network
