+++ mongo-init.js
@@ -1,5 +1,5 @@
-db = db.getSiblingDB('testdb');
+db = db.getSiblingDB('lol');
 db.createCollection('summoners');
-db.createCollection('match');
-db.createCollection('matchID');
+db.createCollection('match_data');
+db.createCollection('match_id');
 console.log("Database and collection initialized successfully!");
